# Conflicts:
#	src/main/java/com/findwork/findwork/Repositories/JobOfferRepository.java

+++ src/main/java/com/findwork/findwork/Controllers/JobOfferController.java
@@ -1,5 +1,6 @@
 package com.findwork.findwork.Controllers;
 
+import com.findwork.findwork.Entities.JobApplication;
 import com.findwork.findwork.Entities.JobOffer;
 import com.findwork.findwork.Entities.Users.UserCompany;
 import com.findwork.findwork.Entities.Users.UserPerson;
@@ -21,14 +22,26 @@ import java.util.UUID;
 
 @Controller
 @AllArgsConstructor
-@RequestMapping("/offer")
+@RequestMapping("/offers")
 public class JobOfferController {
     private final ValidationService validationService;
     private final OfferService offerService;
 
     @GetMapping("/")
-    public String getAllOffers(Model model) {
-        List<JobOffer> offers = offerService.getAllOffers();
+    public String getAllOffers(Model model,
+                               @RequestParam(required = false) String search,
+                               @RequestParam(required = false) String jobCategory,
+                               @RequestParam(required = false) String jobLevel) {
+        List<JobOffer> offers;
+
+        if (jobCategory != null && jobCategory.equals("--Any--"))
+            jobCategory = null;
+
+        if (jobLevel != null && jobLevel.equals("--Any--"))
+            jobLevel = null;
+
+        offers = offerService.getOffers(search, jobCategory, jobLevel);
+
         model.addAttribute("offers", offers);
         model.addAttribute("levels", JobLevel.values());
         model.addAttribute("categories", Category.values());
@@ -36,40 +49,53 @@ public class JobOfferController {
     }
 
     @GetMapping("/create")
-    String getCreateOfferPage(Model model) {
+    String getCreateOfferPage(Model model, Authentication auth) {
+        UserCompany company = (UserCompany) auth.getPrincipal();
         model.addAttribute("levels", JobLevel.values());
         model.addAttribute("categories", Category.values());
+        model.addAttribute("company", company);
         return "createOffer";
     }
 
     @PostMapping("/create")
-    public String createOffer(Authentication auth, CreateJobOfferRequest request, Model model) {
+    public String createOffer(Authentication auth, CreateJobOfferRequest request, RedirectAttributes attr) {
         UserCompany company = (UserCompany) auth.getPrincipal();
         JobOffer questionableOffer;
         try {
             validationService.validateCreateJobOfferRequest(request);
             questionableOffer = offerService.createOffer(request, company);
         } catch (Exception e) {
-            model.addAttribute("error", e.getMessage());
-            return "createOffer";
+            attr.addFlashAttribute("error", e.getMessage());
+            return "redirect:/offers/create";
         }
-        return "redirect:/offer/" + questionableOffer.getId();
+
+        return "redirect:/offers/" + questionableOffer.getId();
     }
 
-    @DeleteMapping("/{id}/remove")
-    public String removeOffer(@PathVariable UUID id, Model model) {
+    @PostMapping("/{id}/remove")
+    public String removeOffer(@PathVariable UUID id, RedirectAttributes attr, Authentication auth) {
+        UserCompany company = (UserCompany) auth.getPrincipal();
+        if (!offerService.loadOfferById(id).getCompany().getId().equals(company.getId()))
+            return "redirect:/offers/" + id;
+
         try {
             offerService.removeOffer(id);
         } catch (Exception e) {
-            model.addAttribute("error", e.getMessage());
-            return "redirect:/offer" + id;
+            attr.addFlashAttribute("error", e.getMessage());
+            return "redirect:/offers/" + id;
         }
-        return "redirect:/";
+
+        return "redirect:/company/" + company.getId();
     }
 
     @GetMapping("/{id}/edit")
-    String getEditOfferPage(@PathVariable UUID id, Model model) {
+    String getEditOfferPage(@PathVariable UUID id, Model model, Authentication auth) {
         JobOffer offer = offerService.loadOfferById(id);
+
+        UserCompany company = (UserCompany) auth.getPrincipal();
+        if (!offer.getCompany().getId().equals(company.getId()))
+            return "redirect:/offers/" + id;
+
         model.addAttribute("offer", offer);
         model.addAttribute("levels", JobLevel.values());
         model.addAttribute("categories", Category.values());
@@ -77,13 +103,38 @@ public class JobOfferController {
     }
 
     @GetMapping("/{id}")
-    String getOfferPage(@PathVariable UUID id, Model model) {
+    String getOfferPage(@PathVariable UUID id, Model model, Authentication auth) {
+        if (auth != null && auth.isAuthenticated() && auth.getPrincipal() instanceof UserPerson) {
+            boolean saved = offerService.checkSaved((UserPerson) auth.getPrincipal(), id);
+            model.addAttribute("saved", saved);
+
+            boolean applied = offerService.checkApplied((UserPerson) auth.getPrincipal(), id);
+            model.addAttribute("applied", applied);
+        }
+
         model.addAttribute("offer", offerService.loadOfferById(id));
         return "offer";
     }
 
+    @GetMapping("/{id}/applications")
+    public String getOfferApplications(@PathVariable UUID id, Model model, Authentication auth) {
+        UserCompany company = (UserCompany) auth.getPrincipal();
+        if (!offerService.loadOfferById(id).getCompany().getId().equals(company.getId()))
+            return "redirect:/offers/" + id;
+
+        List<JobApplication> applications = offerService.getOfferApplications(id);
+
+        model.addAttribute("offer", offerService.loadOfferById(id));
+        model.addAttribute("applications", applications);
+        return "offerApplications";
+    }
+
     @PostMapping("/{id}")
-    public String editOffer(@PathVariable UUID id, EditJobOfferRequest request, Model model) {
+    public String editOffer(@PathVariable UUID id, EditJobOfferRequest request, Model model, Authentication auth) {
+        UserCompany company = (UserCompany) auth.getPrincipal();
+        if (!offerService.loadOfferById(id).getCompany().getId().equals(company.getId()))
+            return "redirect:/offers/" + id;
+
         try {
             validationService.validateEditJobOfferRequest(request);
             offerService.editOffer(id, request);
@@ -91,7 +142,8 @@ public class JobOfferController {
             model.addAttribute("error", e.getMessage());
             return "editOffer";
         }
-        return "redirect:/offer/" + id;
+
+        return "redirect:/offers/" + id;
     }
 
     @PostMapping("/{id}/apply")
@@ -101,12 +153,13 @@ public class JobOfferController {
             validationService.validateUserinfo(user);
             offerService.createApplication(user, id);
         } catch (Exception e) {
+            System.out.println(user.getName());
             atrr.addFlashAttribute("error", e.getMessage());
-            return "redirect:/offer/" + id;
+            return "redirect:/offers/" + id;
         }
 
         atrr.addFlashAttribute("success", "You have applied successfully!");
-        return "redirect:/offer/" + id;
+        return "redirect:/offers/" + id;
     }
 
     @PostMapping("/{id}/cancel")
@@ -116,11 +169,11 @@ public class JobOfferController {
             offerService.deleteApplication(user, id);
         } catch (Exception e) {
             atrr.addFlashAttribute("error", e.getMessage());
-            return "redirect:/offer/" + id;
+            return "redirect:/offers/" + id;
         }
 
         atrr.addFlashAttribute("success", "Canceled job application!");
-        return "redirect:/offer/" + id;
+        return "redirect:/offers/" + id;
     }
 
     @PostMapping("/{id}/save")
@@ -130,11 +183,11 @@ public class JobOfferController {
             offerService.saveOffer(user, id);
         } catch (Exception e) {
             attr.addFlashAttribute("error", e.getMessage());
-            return "redirect:/offer/" + id;
+            return "redirect:/offers/" + id;
         }
 
         attr.addFlashAttribute("success", "Offer saved!");
-        return "redirect:/offer/" + id;
+        return "redirect:/offers/" + id;
     }
 
     @PostMapping("/{id}/unsave")
@@ -144,10 +197,10 @@ public class JobOfferController {
             offerService.unsaveOffer(user, id);
         } catch (Exception e) {
             atrr.addFlashAttribute("error", e.getMessage());
-            return "redirect:/offer/" + id;
+            return "redirect:/offers/" + id;
         }
 
         atrr.addFlashAttribute("success", "Offer removed from saved!");
-        return "redirect:/offer/" + id;
+        return "redirect:/offers/" + id;
     }
 }
